requirements.txt
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pyTelegramBotAPI==4.30.0
python-telegram-bot==22.6
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
